change to use bfloat16

+++ scripts/train_contrast.py
@@ -54,7 +54,7 @@ argParser.add_argument("--save_checkpoint_dir", type=str, default="./checkpoints
 argParser.add_argument("--load_model_checkpoint_path", type=str, default="")
 argParser.add_argument("--load_optimizer_scheduler_checkpoint_path", type=str, default="")
 
-argParser.add_argument("--torch_dtype", type=utils_argparse.str2dtype, default="float16")
+argParser.add_argument("--torch_dtype", type=utils_argparse.str2dtype, default="bfloat16")
 argParser.add_argument("--batch_size_per_device", type=int, default=8)
 argParser.add_argument("--num_epochs", type=int, default=12)
 argParser.add_argument("--save_every_epochs", type=int, default=1)
